Port the app chooser portal
Rename org.freedesktop.portal.AppChooser to org.freedesktop.portal.OpenURI
and redo the implementation along the same lines as the file chooser
portal. We use the new app chooser interface provided by the gtk
backend. More of the work is now done in the frontend: listing suitable
applications, and launching the chosen one. The backend provides a
simple chooser UI.