open-uri: Directly open the URI with the latest used application if possible
If the same application has been repeteadly chosen a few times already,
the frontend will skip calling the backend to open the chooser dialog
and launch the latest used application directly from here.

For now, the threshold is hardcoded (5 times) and there's no way to
reset the situation once gone over the threshold without using the
org.freedesktop.impl.portal.PermissionStore interface directly. Perhaps
a good solution could be to have some kind of integration with gnome
control center (or somewhere else), so that those associations can be
reset/removed, but that's clearly out of the scope of this patch.