Rework the backend API
This is along the same lines as the frontend api, with a shared
Request interface that has Close and Response. The main difference
is that we pass the handle in from the frontend, so backend and
frontend handle share the same id. This lets us avoid a separate
hash table to map backend handles to frontend handles in every
portal implementation.