More work on the print portal
Rename PrintFile to Print.

Add a PreparePrint request that uses a print dialog to send print
settings and page setup back to the sandboxed app, for formatting
the document. It also sends back a token which can be used with
the Print request to have the formatted document printed without
another print dialog. The portal keeps the token valid and remembers
the settings associated with it for a short time.

The Print request is changed to take a file descriptor instead of
a filename.