open-uri: Allow configuring threshold to skip the dialog per application
Add a third field to the data stored to store the threshold used to know
when not to show the dialog for http(s) schemes, so that it can be set
to something different from the hardcoded default value, per application.

Also, organize this information being stored into independent strings,
instead of keeping adding it to the same one, separated with colons.

https://github.com/flatpak/xdg-desktop-portal/issues/32