open-uri: Set sane defaults when reading info from the permissions store
Due to a mistake introduced in commit b8f34737, the data from the permissions
store is not correctly handled the first time a flatpak requires opening an
URI, resulting in setting the threshold for that app always to zero, which
prevents the dialog from ever showing again.

To fix this, set defaults for the values returned by get_latest_choice_info(),
so that it always returns the less restrictive values by default (i.e. no
threshold, always ask) unless a different one has been explicitly set.

https://github.com/flatpak/xdg-desktop-portal/issues/32