open-uri: Implement a more permissive policy for the OpenURI portal
As agreed during the GTK+ hackfest, we are implementing a new policy
for the OpenURI portal to try to improve the user's experience.

As per this new policy, the chooser dialog will NOT be shown for URIs
considered "safe" (Internet URIs, paths to local directories) when either:
  * A default application is set in the system to handle the content type.
  * Only one possible selection would be available to show on the dialog.

This would be an intermediate step while we don't have a better way to
handle default applications for specific schemes and/or content types.

Fixes: https://github.com/flatpak/xdg-desktop-portal/issues/93