Make email portal race narrower
The portal API pattern is inherently racy: The app receives the
response call with the handle of the Request object, but the
Response signal may have already been emitted from the portal
by the time the app has set up a listener for it. Normally, this
should not matter, since the portal api is expected to involve user
interaction - by the time the user is done with a dialog, the
app has certainly established its signal listener.

But, in the case of the email portal, we don't wait for the
'dialog' to be done, we simply launch the email client and
return, which can be rather quick.

To make this race less likely, at least complete the ComposeEmail
call before forwarding it to the backend.