open-uri: Don't crash when opening files skipping the app chooser
When OpenFile is used to open a directory and either there's a default
application to handle file:// URIs, or only one appropriate application
is installed (e.g. Nautilus), we would hit a crash with the current code
since launch_default_for_uri() will be passed a NULL uri parameter.

This fixes that by ensuring that uri is always properly set at that point.

https://github.com/flatpak/xdg-desktop-portal/issues/129