document-portal: reimplement app_has_file_access
This adds back the support for the AS_NEEDED_BY_APP flag, by
shelling out to flatpak info --file-access=PATH APPID, which
is available to later versions of flatpak.

If that fails we fall back to a custom implementation that is a
bit simplified but will catch the most common cases and will
not cause false positives.