Broaden what are valid app ids to allow snap.$PKGNAME
We now allow numbers as the first character in an app id segment.
We also allow names with two components.
This allows us to map any valid snap package name to the
form snap.$PKGNAME.

This also means we could technically create a document for an
that is not a valid flatpak app id. However, flatpaks would just
never look these up, so this is not a problem.