utils: Track container instances using D-Bus Containers1 interface
At the moment we always call GetConnectionCredentials, so we don't need
to make our use of Containers1 conditional, or cope with it failing:
if the container instance object path is in the credentials, then it's
all going to work.

For peers that are on a Containers1 private server, we cache the
app info per container instance as well as per sender
(reference-counting to minimize duplication).

Signed-off-by: Simon McVittie <smcv@collabora.com>