utils: Get D-Bus container instance from header field if possible
In the case where Containers1.RequestHeader() succeeds, we can use
a fast-path by relying on the dbus-daemon adding a trusted container
instance path to the message header.

This is currently ifdef'd out because it slightly abuses the GLib
API by passing an arbitrary integer constant to
g_dbus_message_get_header() where a member of GDBusMessageHeaderField
is expected. Allison Lortie says this isn't really meant to work and is
undefined behaviour, although in practice it does work well enough for
testing. We shouldn't really rely on it until support for the new header
field is added to GLib, so for now, build with
-DRELY_ON_GLIB_IMPLEMENTATION_DETAILS (at your own risk) if you want to
test this code path.

Signed-off-by: Simon McVittie <smcv@collabora.com>