Don't register http: uris
When we have a sandboxed handler, we try to register
the uri with the document portal to ensure that the
handler can access it. But this only makes sense if
the uri is a file: one.

This was causing failures when opening http: uris
via the portal with a sandboxed epiphany.

See https://github.com/flatpak/xdg-desktop-portal/issues/173