document-portal: Support non-O_PATH fds for adding existing files
As discussed in https://github.com/flatpak/xdg-desktop-portal/issues/167
using O_PATH to pass proof of access to a file is not good enough for
e.g. snap, as it only means the sandbox can see the pathname, not that
it can e.g. open it. So, if the fd is not O_PATH we use the open
flags to detect if the file should be writable or not.

For directories we still use O_PATH because there is no way to prove
writability to a directory. However, AddNamed*() is only callable from
the host, so that is not really a problem.

We still support O_PATH for host and flatpak callers, because those
are safe and there are still code out there using this.