OpenURI: Fix misleading warning when accessing the permission store
The warning message is wrong since we're not updating the permission
store, but accessing it, at that point so we need to fix that. Besides
the warning message should only shown when the error is other than
NOT_FOUND, since not having a entry recorded in the desktop-apps table
for a particular content-type is a perfectly fine scenario.