OpenURI: Don't skip the dialog based on content type if a threshold is set
If a threshold is set for a given content-type/schema and app ID in the
permission store, we should make the decision on whether to show the
app chooser dialog or not based only on whether such threshold has been
reached by consistently selecting the same app a big enough amount of
consecutive times, not based on whether we're opening a directory or
a HTTP/HTTPs URI (only cases when we currently might skip the dialog).

If a threshold is not set (this is the default unless the permission store
has been altered with an external tool), the current behaviour is kept
and the dialog will still be skipped for those special cases.

https://github.com/flatpak/xdg-desktop-portal/issues/149