document-portal: Fix XdpInode leak
When creating nodes we forgot to free them, so inodes would never
go away, even when the kernel told us it forget about them.

This causes issues when re-creating forgotten files, such as #195

This hits the TODO in xdp_inode_create_file():

      if (truncate)
        {
          /* TODO: Handle extra truncate for existing file */
          errno = ENOSYS;

We should really fix the TODO, but with the proper inode refcounting
we avoid hitting it in this case, because the first lookup of the now
deleted file will cause the kernel to forget() the inode, which
deletes the pre-existing XdpInode.

Fixes #195