network-monitor: Use ::network-changed for change notifications
The ::network-changed signal is emitted when the network configuration
changes, which doesn't necessarily mean a change of network availability
or connectivity. For example connecting to a VPN doesn't usually change
the general online state, but as additional services become available,
clients may want to retry previously failed connection attempts.

To support this use case in the portal implementation, simply relay
the "native" network monitor's ::network-changed signal instead of
listening for property changes.