document-portal: Ensure we get a clean unmount in tests
In the tests, we set TEST_DOCUMENT_PORTAL_FUSE_STATUS to a path in the
environment, which triggers the fuse filesystem to write a status
to this file. If it is not 'ok', the unmount is considered unclean
and we fail the tests.

For now, clean means that we don't leak any XdpInodes, which means
we also have to unref all global refs after unmount.