document-portal: Fix a leak of an XdpInode
When a temporary (i.e. not document) file is created in a document
dir, it gets an extra ref (until unlink) to keep the filename alive.
However, when the temp file gets promoted to a document file due
to being rename to the real basename we forgot to remove this
extra ref.