document-portal: Don't cache entries and attributes for temp files
Even the temp files are not really under our control and could be
modified outside of via fuse.

In addition, a tempfile can be promoted to a regular file with a
rename, and the kernel will still keep the old cached data for the
tmpfiles. This is a problem for example when a document file was
created in such a way, and then it gets removed from the host side.
In this case the kernel thinks the old file is still there and
an open (O_CREAT) will fail because it does an open() instead
of a create().