notification: Drop suspicious icons
Do a more thorough icon validation. We reject
anything that is not a serialized themed icon or
bytes icon right away. Later, when we are in a
thread, we run do a second validation step where
we spawn the icon validator utility. If this
fails, we drop the icon from the notification.

For good measure, we isolate the icon validation
code in a bubblewrap sandbox, so this commit adds
a dependency on bubblewrap.