notification: Write back permissions
We want to use the fact that there are permissions in the
permission store for an app to hide or show the corresponding
row in the control-center. Basically, we want the permission
store to record which apps have used this portal. So, make
the notification portal write back a "yes" when the app is
using it.