screen-cast: Add cursor modes
Used to let the portal user decide whether to hide, embed or side
channel the cursor sprite of a screen cast.

Hiding it means it won't be visible. Embedding it means it'll be
embedded into the screen cast video stream buffers. Side channelling it
means it will be sent as PipeWire stream metadata, leaving the stream
video content itself free of cursor sprites.