utils: add function to map process ids
Containers can use pid_namespaces(7) to encapsulate the processes
running inside the container from seeing the process ids (pids)
of the host. A generic way to translate a process identifier from
within the container to the outside that does not need the
cooperation of the target process is to use /proc which contains
the information about the namespace for each process as well as
the process identifier of the process in each namespace. The
namespace id can be obtained via '/proc/<pid>/ns/pid'; the pids
for all the namespaces via '/proc/<pid>/status' and its 'NSpid'
field.
For Flatpaks, which uses bubble wrap, the bwrapinfo.js file has the
child-pid field, which is the bwrap process inside the sandbox.
Via this pid the pid namespace id can be determined. All process in
/proc are then scanned and filtered for ones that belong to the same
pid namespace. For those their 'status' file is parsed and the pid
within the namespace matched against the ones that need to be
translated.