screen-cast: Set pipewire.access.portal.* properties
Starting with version 0.2.6, the PipeWire daemon expects certain
properties to be set when opened by the portal, specifically
pipewire.access.portal.app_id, set to the application ID or an empty
string if unavailable, and pipewire.access.portal.media_roles, set to a
comma seperated list of media roles the PipeWire daemon should query the
portal for permissions.  Media roles not present in the role list keep
their permission set by the portal before handing it over to the
application.

For the screen cast, no media roles are to be queried, as permissions
are setup up front.