background: Monitor running apps
Implement a facility that monitors running flatpaks,
notifies the user if they run in the background, and
kills them when they are not allowed to run in the
background.

The platform-dependent parts of this (getting the
state of running applications, and notifying the user)
are handled by a portal backend.