Add camera portal
The responsibility of the camera portal is to query the Access API
whether the application should be allowed to open camera devices or not.
For applications allowed, it can open a PipeWire remote with permissions
initially set to "denied", with PipeWire itself querying the permission
store for up to date camera permissions.