gamemode: add support for gamemode 1.4 APIs
GameMode 1.4 added a set of new APIs that mirror the existing ones
but where once process can act on behalf of the another. The only
difference of the new function calls to the existing one is the
additional "requester" argument (a process id). In case of faltpak
this new pid argument also needs to be to be mapped from the pid
namespace of the sandbox to the pid namespace of the host.