Add a secret portal
This adds a portal interface to provide sandboxed applications with
master secret to locally encrypt credentials. It works as follows: the
application first creates a pipe, sends the write end to the portal,
and the portal sends back a secret value through the FD. That way, we
can avoid any transport encryption.

If the master secret doesn't exist in keyring, the portal generates a
random secret. This step needs to be done with a standard conforming
crypto library as the secret is used for encryption. Therefore the
backend implementation resides in gnome-keyring.