gamemode: new D-Bus method based on pidfds
Add new pidfd based APIs, similar to the existing *PID ones. But
instead working with process ids it works with pidfds, obtained
via pidfd_open(2). These can then be used to obtain the process
ids (pids) for the corresponding processes outside the container,
without the need of the complicated mapping.
The portal itself is just forwarding the calls. If the actual
GameMode daemon is not yet implementing the calls, they will fail
with "UnknownMethod" and the client implementation should handle
that gracefully, falling back to the non-pidfd based methods.
The portal will verify that the supplied fds are indeed pidfds
that can be resolved to a valid pid in the portals pid namespace.