open-uri: Revamp the policy
Collect all the code that handles the 'shall we open a dialog'
policy in one place. And tweak the rules we apply slightly, with
the goal of being less annoying by default, while still supporting
paranoid use cases.

By default, we skip the app chooser and directly launch the
default handler for http and inode/directory requests. We also
skip the app chooser if there is only one choice.

Lastly, we skip the app chooser if the last choice he been
selected often enough by the user (the threshold we use here
is 3).

The application can override these defaults by passing the
'ask' open with the OpenURI request. If that option is present,
we always show the app chooser.

The user can override everything by setting the threshold store
in the permission store to the special value 2147483647 (INT_MAX).
If the threshold has this value, we always show the app chooser.