openuri: Allow a per-type always-ask option
Optionally read per-entry data to determine if
we should unconditionally present the app chooser.

This allows the user to say 'always ask for http',
regardless of what handler is being used.

The format of the data is an a{sv}, and we look
for the 'always-ask' key, expecting a boolean
value.