document-portal: Fuse, try to keep down nr of open fds
To avoid wasting fds we queue a dentry invalidate after every
successful lookup operation inside a document dir. This will
cause all outstanding dentries to be freed, thereby making the
kernel emit FORGET for inodes not used other than in the dcache.

This makes the stress-test go from max ~4000 outstanding fds to ~1000
for me, which should be enough to pass on systems with a hard limit
of 4096 fds.