document-portal: Don't invalidate forgotten inodes
We only want to invalidate the entries if the kernel didn't already
tell us to foget the inode, so we keep the parent inodes of all
domains alive so we can check their kernel ref counts. These parents
are all virtual and few so keeping them alive longer doesn't use a lot
of memory or file descriptors.

Without this i was sometimes getting ENOTDIR when telling the kernel
to invalidate an entry, presumably because the parent inode I passed
had been forgotten and reused for another file.