document-portal: Fix volatile app ids in fuse test
We do verify that just looking for a valid app id directory creates
it, because that is how the mount of them works in flatpak. However
such directories are not persistent in the database, so if the kernel
forgets the inode we will not see it again in readdir().

This failed the test which assumed the created app id would live forever.
Fix this by marking it volatile and not failing it if goes away.