document-portal: Ensure root docdir inode alive while there are other inodes
The root inode of the document directory is cheap to keep alive (no fd)
and it keeps the document domain alive. If the kernel forgets it and
it has no other reference then the next time the kernel asks for it
we will create a new document domain. This is bad because other existing
inodes might still be using the old document domain.

So, we keep a ref from the non-root inodes of a domain to the root inode.