document-portal: Make Temp files keep the inode around
We need the temp file to keep the inode alive because otherwise
the root inode can become forgotten which leads to the domain
tempname mapping to be lost and the tempfile disappears.

This is a cyclic ref:
 root inode -> domain -> tempfile dict -> tempfile -> inode -> root inode

But it is explicitly broken when the temp file is unlinked (which
causes the tempfile dict -> tempfile ref to go away.