document-portal: Try to make inode nrs persistent
Try to generate inode nrs based on the underlying files device and inode nr,
but hashing them (combined with app/doc id to avoid same ino in different
domains).

This means that unless we run into some conflict, we will not report
a different inode nr for the same file just because the kernel forgot
the inode inbetween two lookups.