open-uri: Show app chooser when default app does not exist
Fix a possible crash when the application that would be selected to
avoid the app chooser is not available (anymore).

This would have happened when eog.desktop got renamed to
org.gnome.eog.desktop but was launched enough times as eog.desktop to
make it the default.

Based on https://github.com/flatpak/xdg-desktop-portal/pull/481