permission-store: Add GetPermission
We currently have SetPermission and DeletePermission which operate
with specific apps, but we didn't have a GetPermission method that
does the same. Instead, we rely on LookUp which brings all (table,
id, app) tuples.

From what I have seen,in GNOME Control Center and what I needed
to do for Flatseal, we call LookUp, check if the app is present
among the retrieved tuples,  to then access the permission data
we are looking for (e.g, ["yes"], etc).

As more UI tools ala-flatseal appear, a GetPermission would make
more sense for such tools as these operate per-app basis.

Another reason is that the permission store data tends to grow,
which means that  LookUp  will keep bringing more and more data
unnecessarily for the case described.

Closes #583