utils: Add fallback definition for G_DBUS_METHOD_INVOCATION_HANDLED
When implementing GDBus methods, it's not always obvious that
"return TRUE" on error is correct - but in fact it is, because the
boolean result is handled/unhandled rather than the usual success/error.
GLib 2.68 introduced macros to make this more obvious. Backport them
here for older GLib versions.

I'm not systematically converting existing code right now, but the new
macros can be used in new code.

Signed-off-by: Simon McVittie <smcv@collabora.com>