Add a portal for Emulated Input
This is a portal with an impl backend, with two calls each:
CreateSession which determines if a client may emulate input. That call
triggers the permissions check in the portal and the interactive
permission access.

The second call is a simple connection call to the EIS implementation
which should return immediately and provides the fd to talk to the EIS
implementation.

The implementation backend is required because there may be multiple EIS
implementations running, so we rely on the backend to know which one we
actually need to connect to.