document-portal-fuse: Ensure that session is never destroyed while in use
The fuse session (that is owned by the fuse thread) may be used from
the main thread in some cases, but this may happen also just before the
owner thread is about to destroy and nullify it, possibly causing to
use invalid data.

To avoid this, use a mutex to protect the access to the session that always
ensures that nobody else is currently using it.
In this way, while is owned by the main (or another) thread, the fuse
thread can't unset and destroy it until the owner has not done with it.