test-portals: Terminate the test processes via SIGTERM
We used to just kill the test processes without caring whether they were
successful or not.

However in all our tests we assume that the daemons are well behaving
and so that they're not crashing and correctly reacting to SIGTERM.
Sadly we didn't really check this, so if a daemon was misbehaving in on
shutdown (like in the case of a race condition leading to a deadlock) we
were just ignoring this.

To avoid it, let's manually simulate stopping the processes gracefully
via SIGTERM and check that all the processes just terminated because of
it or that if that is handled, the process has successfully exited.

This may be relevant in case a subprocess just fails or crashed while
handling the termination.