Make x-d-p start on session start
We need x-d-p to start on session start, since it will be responsible
for migrating dynamic launchers for apps that have been renamed, e.g.
via Flatpak's end-of-life-rebase mechanism. This is important because
the app may be installed system-wide and shared by multiple users,
whereas the dynamic launchers are per-user, so when Flatpak applies an
update that renames an app, it can only be responsible for migrating the
launchers for the current user to the new app ID. Other users' launchers
are renamed by xdg-desktop-portal when they log in to a graphical
session.