Bring back in-house icon validator
This reverts commit 2fb498f92f8e1c41049dedf887690083fa12f6d1 and makes
related changes so that we use our own copy of Flatpak's icon validator
(taken from icon-validator/validate-icon.c in Flatpak 1.13.1).

Here are the reasons for bringing back this copied code:
(1) To avoid a circular dependency between x-d-p and Flatpak. Debian has
been patching this code back in for years for this purpose.
(2) We don't want icon validation to be optional since it's a security
measure. Having our own binary makes it easier to make it required.
(3) For the new dynamic launcher portal, we need to determine the size
and format of icons. The validator already gathers this information so
it makes sense to just have it print those things to stdout. This means
our copy of this code will diverge from Flatpak.
(4) It is too small to be worth maintaining as a separate repo.