doc: fix links and some formatting in the portal.* interfaces
Fix the various links to portals, methods and signals, with the
resulting re-formatting required.

Fix some of the arguments that previously spanned over two lines, this
results in incorrect rendering.

Fix the Wallpaper value lists to use a proper variablelist like the
other portals.

And smuggle in a typo fix or two and some whitespace fixes vim decided
to apply.