Allow portal access non-flatpak sandboxes
Currently, the portal tries to check the process filesystem to inspect
its flatpak metadata, but this is not possible for non-flatpak
sandboxes, for which D-Bus access is out-of-scope here.

In particular, this affects firejail applications which have been
granted full d-bus access (e.g.: `d-feet`).

Fixes https://github.com/flatpak/xdg-desktop-portal/issues/737