Change generated D-Bus code namespace to XdpDbus(Impl)
Tests compile xdp-dbus.c and xdp-impl-dbus.c, and also link to
libportal. This, as is, isn't possible without symbol conflicts, since
xdp-dbus.c share the same symbol namespace with libportal, meaning it's
not possible to use any types that have the same name, e.g. XdpSession,
as it's likely that the type instance from xdp-dbus.c will be used
instead of the one from libportal.

Fix this by changing the namespace of code generated with gdbus-codegen
to XdpDbus and XdpDbusImpl to avoid these conflicts. The Impl part is
changed too to make things consistent.